refactor: define const values to constants.ts
